refactors to use modules
